[PO-7] refactored POMs
