+++ poster-dao/src/main/java/org/pti/poster/dao/DatabaseClient.java
@@ -1,4 +1,4 @@
-package org.pti.dao;
+package org.pti.poster.dao;
 
 
 import com.mongodb.MongoClient;
