+++ poster-integration-tests/pom.xml
@@ -9,7 +9,6 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<artifactId>poster-integration-tests</artifactId>
-	<dependencies>
-	</dependencies>
 	<name>poster-integration-tests</name>
+
 </project>
