+++ poster-model/pom.xml
@@ -10,16 +10,18 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<artifactId>poster-model</artifactId>
+	<name>poster-model</name>
+
 	<dependencies>
-		<dependency>
-			<groupId>org.projectlombok</groupId>
-			<artifactId>lombok</artifactId>
-		</dependency>
-        
         <dependency>
             <groupId>org.springframework.data</groupId>
             <artifactId>spring-data-mongodb</artifactId>
         </dependency>
+
+		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+		</dependency>
 	</dependencies>
-	<name>poster-model</name>
+
 </project>
