+++ poster-service/pom.xml
@@ -10,27 +10,23 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>poster-service</artifactId>
+    <name>poster-service</name>
 
     <dependencies>
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-core</artifactId>
+            <groupId>org.pti</groupId>
+            <artifactId>poster-model</artifactId>
+            <version>${project.version}</version>
         </dependency>
 
         <dependency>
-            <groupId>com.mangofactory</groupId>
-            <artifactId>swagger-springmvc</artifactId>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-core</artifactId>
         </dependency>
         
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.pti</groupId>
-            <artifactId>poster-model</artifactId>
-            <version>${project.version}</version>
         </dependency>
 
         <dependency>
@@ -42,7 +38,11 @@
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
         </dependency>
+
+        <dependency>
+            <groupId>com.mangofactory</groupId>
+            <artifactId>swagger-springmvc</artifactId>
+        </dependency>
     </dependencies>
 
-    <name>poster-service</name>
 </project>
