+++ poster-service/src/main/java/org.pti.poster.service/PersonRespository.java
@@ -1,4 +1,4 @@
-package org.pti.poster;
+package org.pti.poster.service;
 
 
 import org.pti.poster.model.Person;
