+++ poster-service/src/main/java/org.pti.poster.service/SpringApplicationConfiguration.java
@@ -1,4 +1,4 @@
-package org.pti.poster;
+package org.pti.poster.service;
 
 
 import com.mangofactory.swagger.plugin.EnableSwagger;
