+++ poster-spring-boot/src/main/java/org/pti/poster/boot/Application.java
@@ -1,7 +1,7 @@
 package org.pti.poster.boot;
 
 
-import org.pti.poster.SpringApplicationConfiguration;
+import org.pti.poster.service.SpringApplicationConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
