+++ poster-spring-boot/src/test/java/org/pti/poster/boot/HelloControllerTest.java
@@ -7,7 +7,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.pti.poster.controllers.HelloController;
+import org.pti.poster.boot.controllers.HelloController;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockServletContext;
